chore: put some oidc test constants in separate file
