+++ src/app/modules/spcp/__tests__/spcp.oidc.client.spec.ts
@@ -30,15 +30,7 @@ import {
 } from '../spcp.oidc.client.types'
 
 import {
-  CP_OIDC_NDI_DISCOVERY_ENDPOINT,
-  CP_OIDC_NDI_JWKS_ENDPOINT,
-  CP_OIDC_RP_CLIENT_ID,
-  CP_OIDC_RP_REDIRECT_URL,
   cpOidcClientConfig,
-  SP_OIDC_NDI_DISCOVERY_ENDPOINT,
-  SP_OIDC_NDI_JWKS_ENDPOINT,
-  SP_OIDC_RP_CLIENT_ID,
-  SP_OIDC_RP_REDIRECT_URL,
   spOidcClientConfig,
   TEST_CP_RP_PUBLIC_JWKS,
   TEST_CP_RP_SECRET_JWKS,
@@ -46,6 +38,16 @@ import {
   TEST_NDI_SECRET_JWKS,
   TEST_SP_RP_PUBLIC_JWKS,
   TEST_SP_RP_SECRET_JWKS,
+} from './spcp.oidc.test.constants'
+import {
+  CP_OIDC_NDI_DISCOVERY_ENDPOINT,
+  CP_OIDC_NDI_JWKS_ENDPOINT,
+  CP_OIDC_RP_CLIENT_ID,
+  CP_OIDC_RP_REDIRECT_URL,
+  SP_OIDC_NDI_DISCOVERY_ENDPOINT,
+  SP_OIDC_NDI_JWKS_ENDPOINT,
+  SP_OIDC_RP_CLIENT_ID,
+  SP_OIDC_RP_REDIRECT_URL,
 } from './spcp.test.constants'
 
 jest.mock('openid-client')
