+++ src/app/modules/spcp/__tests__/spcp.test.constants.ts
@@ -1,17 +1,11 @@
 import { MyInfoMode } from '@opengovsg/myinfo-gov-client'
 import { ObjectId } from 'bson'
 import crypto from 'crypto'
-import fs from 'fs'
 import _ from 'lodash'
 
 import { ISpcpMyInfo } from 'src/app/config/features/spcp-myinfo.config'
 import { ILoginSchema, IPopulatedForm } from 'src/types'
 
-import {
-  PublicJwks,
-  SecretJwks,
-  SpcpOidcClientConstructorParams,
-} from '../spcp.oidc.client.types'
 import { JwtName } from '../spcp.types'
 
 export const MOCK_SERVICE_PARAMS: ISpcpMyInfo = {
@@ -200,28 +194,6 @@ export const MOCK_CP_OIDC_JWT_PAYLOAD = {
   rememberMe: true,
 }
 
-export const TEST_SP_RP_PUBLIC_JWKS: PublicJwks = JSON.parse(
-  fs.readFileSync('tests/certs/test_sp_rp_public_jwks.json').toString(),
-)
-export const TEST_SP_RP_SECRET_JWKS: SecretJwks = JSON.parse(
-  fs.readFileSync('tests/certs/test_sp_rp_secret_jwks.json').toString(),
-)
-
-export const TEST_CP_RP_PUBLIC_JWKS: PublicJwks = JSON.parse(
-  fs.readFileSync('tests/certs/test_cp_rp_public_jwks.json').toString(),
-)
-export const TEST_CP_RP_SECRET_JWKS: SecretJwks = JSON.parse(
-  fs.readFileSync('tests/certs/test_cp_rp_secret_jwks.json').toString(),
-)
-
-export const TEST_NDI_SECRET_JWKS: PublicJwks = JSON.parse(
-  fs.readFileSync('tests/certs/test_ndi_secret_jwks.json').toString(),
-)
-
-export const TEST_NDI_PUBLIC_JWKS: PublicJwks = JSON.parse(
-  fs.readFileSync('tests/certs/test_ndi_public_jwks.json').toString(),
-)
-
 export const SP_OIDC_NDI_DISCOVERY_ENDPOINT = 'spOidcNdiDiscoveryEndpoint'
 export const SP_OIDC_NDI_JWKS_ENDPOINT = 'spOidcNdiJwksEndpoint'
 export const SP_OIDC_RP_CLIENT_ID = 'spOidcRpClientId'
@@ -231,21 +203,3 @@ export const CP_OIDC_NDI_DISCOVERY_ENDPOINT = 'cpOidcNdiDiscoveryEndpoint'
 export const CP_OIDC_NDI_JWKS_ENDPOINT = 'cpOidcNdiJwksEndpoint'
 export const CP_OIDC_RP_CLIENT_ID = 'cpOidcRpClientId'
 export const CP_OIDC_RP_REDIRECT_URL = 'cpOidcRpRedirectUrl'
-
-export const spOidcClientConfig: SpcpOidcClientConstructorParams = {
-  ndiDiscoveryEndpoint: SP_OIDC_NDI_DISCOVERY_ENDPOINT,
-  ndiJwksEndpoint: SP_OIDC_NDI_JWKS_ENDPOINT,
-  rpClientId: SP_OIDC_RP_CLIENT_ID,
-  rpRedirectUrl: SP_OIDC_RP_REDIRECT_URL,
-  rpSecretJwks: TEST_SP_RP_SECRET_JWKS,
-  rpPublicJwks: TEST_SP_RP_PUBLIC_JWKS,
-}
-
-export const cpOidcClientConfig: SpcpOidcClientConstructorParams = {
-  ndiDiscoveryEndpoint: CP_OIDC_NDI_DISCOVERY_ENDPOINT,
-  ndiJwksEndpoint: CP_OIDC_NDI_JWKS_ENDPOINT,
-  rpClientId: CP_OIDC_RP_CLIENT_ID,
-  rpRedirectUrl: CP_OIDC_RP_REDIRECT_URL,
-  rpSecretJwks: TEST_CP_RP_SECRET_JWKS,
-  rpPublicJwks: TEST_CP_RP_PUBLIC_JWKS,
-}
